requirements.txt
django-bootstrap-v5
fuzzy_match
python-dotenv